Doubleclicking slide listing doesn't set slide title
The codepath to set the document title was only followed if
`displayCurrentSlide` was invoked, which was specifically disabled in
the dblclick event listener. If the slide is selected from the list with
an enter, per the keydown event listener, this codepath is followed.

Solution: abstract out the document.title changing into it's own method,
and call this from both displayCurrentSlide and switchToFull